+++ packages/dear-imgui/ts/src/lib/components/ReactImgui.tsx
@@ -8,7 +8,7 @@ import ReactNativePrivateInterface from "../react-native/ReactNativePrivateInter
 import getWasmModule from "../wasm/reactDearImgui.mjs";
 import { MainModule, WasmExitStatus } from "../wasm/wasm-app-types";
 import { ReactNativeWrapper } from "../components/ReactNativeWrapper";
-import { Primitive } from "../components/ReactImgui/types";
+import { useDearImguiWasm } from "../hooks";
 
 // getWasmModule.preRun = () => {
 //     ENV.MY_FILE_ROOT = "/usr/lib/test";
@@ -29,79 +29,9 @@ export const MainComponent: React.ComponentType<MainComponentProps> = ({
     const isWasmModuleLoading = useRef(false);
     const [wasmModule, setWasmModule] = useState<MainModule | undefined>();
 
-    const onTextChange = useCallback(
-        (id: string, value: string) => {
-            const rootNodeID = id;
-            const topLevelType = "onChange";
-            const nativeEventParam = { value };
-
-            ReactNativePrivateInterface.nativeFabricUIManager.dispatchEvent(
-                rootNodeID,
-                topLevelType,
-                nativeEventParam,
-            );
-        },
-        [wasmModule],
-    );
-
-    const onComboChange = useCallback((id: string, value: number) => {
-        const rootNodeID = id;
-        const topLevelType = "onChange";
-        const nativeEventParam = { value };
-
-        ReactNativePrivateInterface.nativeFabricUIManager.dispatchEvent(
-            rootNodeID,
-            topLevelType,
-            nativeEventParam,
-        );
-    }, []);
-
-    const onNumericValueChange = useCallback((id: string, value: number) => {
-        const rootNodeID = id;
-        const topLevelType = "onChange";
-        const nativeEventParam = { value };
-
-        ReactNativePrivateInterface.nativeFabricUIManager.dispatchEvent(
-            rootNodeID,
-            topLevelType,
-            nativeEventParam,
-        );
-    }, []);
-
-    const onMultiValueChange = useCallback((id: string, values: Primitive[]) => {
-        const rootNodeID = id;
-        const topLevelType = "onChange";
-        const nativeEventParam = { values };
-
-        ReactNativePrivateInterface.nativeFabricUIManager.dispatchEvent(
-            rootNodeID,
-            topLevelType,
-            nativeEventParam,
-        );
-    }, []);
-
-    const onBooleanValueChange = useCallback((id: string, value: boolean) => {
-        const rootNodeID = id;
-        const topLevelType = "onChange";
-        const nativeEventParam = { value };
-
-        ReactNativePrivateInterface.nativeFabricUIManager.dispatchEvent(
-            rootNodeID,
-            topLevelType,
-            nativeEventParam,
-        );
-    }, []);
-
-    const onClick = useCallback((id: string) => {
-        const rootNodeID = id;
-        const topLevelType = "onClick";
-
-        ReactNativePrivateInterface.nativeFabricUIManager.dispatchEvent(rootNodeID, topLevelType, {
-            value: "clicked",
-        });
-    }, []);
-
     const canvasId = useMemo(() => `canvas-${uuidv4()}`, []);
+
+    const { eventHandlers } = useDearImguiWasm(ReactNativePrivateInterface);
 
     useEffect(() => {
         if (canvasRef.current && !isWasmModuleLoading.current) {
@@ -118,14 +48,7 @@ export const MainComponent: React.ComponentType<MainComponentProps> = ({
 
                     //     return `${scriptDirectory}/lib/wasm/${path}`;
                     // },
-                    eventHandlers: {
-                        onTextChange,
-                        onComboChange,
-                        onNumericValueChange,
-                        onMultiValueChange,
-                        onBooleanValueChange,
-                        onClick,
-                    },
+                    eventHandlers,
                 };
 
                 try {
@@ -153,16 +76,7 @@ export const MainComponent: React.ComponentType<MainComponentProps> = ({
         } else {
             return () => {};
         }
-    }, [
-        canvasId,
-        canvasRef,
-        onTextChange,
-        onComboChange,
-        onNumericValueChange,
-        onBooleanValueChange,
-        onMultiValueChange,
-        onClick,
-    ]);
+    }, [canvasId, canvasRef]);
 
     useEffect(() => {
         if (wasmModule) {
