+++ packages/dear-imgui/ts/src/lib/hooks/index.ts
@@ -1,4 +1,5 @@
 import { useWidgetEventManagement } from "./useWidgetEventManagement";
 import { useWidgetRegistrationService } from "./useWidgetRegistrationService";
+import { useDearImguiWasm } from "./useDearImguiWasm";
 
-export { useWidgetEventManagement, useWidgetRegistrationService };
+export { useWidgetEventManagement, useWidgetRegistrationService, useDearImguiWasm };
