+++ packages/dear-imgui/ts/src/lib/index.ts
@@ -1,6 +1,6 @@
 import { ReactImgui } from "./components/ReactImgui/components";
 import { ReactNativeWrapper } from "./components/ReactNativeWrapper";
-import { useWidgetEventManagement, useWidgetRegistrationService } from "./hooks";
+import { useWidgetEventManagement, useWidgetRegistrationService, useDearImguiWasm } from "./hooks";
 import { WidgetRegistrationServiceContext } from "./contexts";
 
 export {
@@ -8,5 +8,6 @@ export {
     ReactNativeWrapper,
     useWidgetEventManagement,
     useWidgetRegistrationService,
+    useDearImguiWasm,
     WidgetRegistrationServiceContext,
 };
